requirements.txt
aiomcache==0.8.2
aioredis==2.0.1
alembic==1.13.2
alembic-postgresql-enum==1.3.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
billiard==4.2.0
cachelib==0.13.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
flower==2.0.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.1
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
phonenumbers==8.13.43
pillow==10.4.0
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.8.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
rich==13.7.1
rsa==4.9
sentry-sdk==2.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
starlette==0.38.2
time-machine==2.15.0
tornado==6.4.1
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
utils==1.0.2
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
WTForms==3.1.2
